refactor: Change names
